requirements.txt
requests
python-dotenv
openai
beautifulsoup4
discord.py